requirements.txt
streamlit
mistralaivenv